requirements.txt
langchain
langchain-community
langchain-core
langchain_openai
python-dotenv
streamlit
huggingface-hub
google-cloud-translate
bs4
faiss-cpu
langchain-cohere
PyPDF2